Optimized the fair spawning system's calculation algorithms and resolved issues with new piglins.
The fair spawning system was doing redundant checks by iterating through the player's entire inventory even when it is not necessary to do so. These inventory / player attribute checks are instead moved in its own function under the spawn weight class.

The spawn weight's calculate raw threat score function no longer accepts a double (which is the total value for a player's attribute this weight is looking for). It instead accepts a player as the parameter and runs a new function, getAttributeValue (player) to find the player's attribute value specified by this weight.

Piglin captains now have a regen buff and royal piglin hunter no longer instantly dies upon spawning.