Mob Skill tasks are no longer ran in their own individual runnable instances.
Mob Skill tasks are now ran in an iterative loop instead of being handled by Bukkit's schedulers. This is done to reduce the total amount of active runnables in this server.

However, there is currently a bottleneck occuring in the skill's run function, as the skill would need to invoke the server's getNearbyEntities function every time they want to use the skill. I was told this is an expensive operation and may be contributing to TPS loss.

This is by means not an optimized solution, but with this new MobSkillRunner running all custom monster tasks iteratively, we can start limiting the total number of skills that can be used every second to save server resources. This will involve creating a new variable in the config file and seperating the update and run functions for the skill