Moved all canUseSkill checks that only invlove the caster down to the canUseSkill function.
This is because teleport skill checks are expensive, as the search radius of these skills are much larger than attack skills. Moving every can use skill check only involving the caster to the canUseSkill function prevents the server from repeatedly running potentially expensive getNearbyEntity calls on situations where the caster is clearly not elligable to use the skill.