Optimized chunk loading algorithm
Instead of grabbing the entities of each individual loaded chunk, it instead grabs all of the worlds the chunks reside at, gets all entities for each world, filters out the entities who are not in the loaded chunks, and then processes them.

This is because chunk.getEntities already does the above process every time I call this function, which has a great negative performance impact on the server, as that process is expensive.